Изменил Task04 в HWT_06

+++ HWT_06/Task04/ILocatable.cs
@@ -7,6 +7,6 @@ namespace Task04
     interface ILocatable
     {
         int X { get; set; }
-        double Y { get; set; }
+        int Y { get; set; }
     }
 }
