+++ HWT_06/Task04/Monster.cs
@@ -8,7 +8,7 @@ namespace Task04
     {
         public abstract string Output { get; set; }
         public abstract int X { get; set; }
-        public abstract double Y { get; set; }
+        public abstract int Y { get; set; }
 
         public abstract void Move(int x, int y);
         public abstract void Attack(Player p);
